Create Zero-Coupon-Bond.hs
This code is defining a smart contract for a zero coupon bond. It defines two parties, investor and issuer, as well as two timeouts, initialExchange and maturityExchangeTimeout.

The initialExchange timeout represents the deadline by which the investor must deposit the discounted price of the bond to the contract. If the investor does not deposit the funds within this time period, the contract will close and the funds will be returned to the investor.

Once the investor has deposited the funds, the issuer has until the maturityExchangeTimeout to pay back the notional price of the bond, which is the discounted price plus any interest that has accrued. If the issuer does not pay back the notional price within this time period, the contract will close and the funds will be returned to the investor.

If the issuer does pay back the notional price within the maturityExchangeTimeout, the contract will transfer the funds to the issuer and then close.