Create CFD.hs
This code defines a contract for differences (CFD) between a party and a counterparty. The contract specifies that both the party and the counterparty will make a deposit in ada, and an oracle will provide the prices of a particular asset at the beginning and at the end of the contract. If the price at the end is higher than the price at the beginning, the party will pay the counterparty the difference between the two prices, up to the amount of the counterparty's deposit. If the price at the end is lower, the counterparty will pay the party the difference, up to the amount of the party's deposit. If the prices are the same, the contract will be closed and both parties will be refunded their deposits.