Create CFD-with-Oracle.hs
To get the price of the asset, you can use the ChoiceValue function and pass in the ChoiceId for the price of the asset. This will give you the value chosen by the oracle for the price of the asset at a given time. You can then use this value in your contract in the same way you would use any other value. For example, you could use it to calculate the difference in price between two points in time and use that to determine how much the parties should pay or receive.